requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
attrs==23.2.0
certifi==2024.6.2
charset-normalizer==3.3.2
coverage==7.5.3
distro==1.9.0
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
frozenlist==1.4.1
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
multidict==6.0.5
openai==1.31.1
packaging==24.0
psycopg2==2.9.9
pydantic==2.7.3
pydantic_core==2.18.4
PyJWT==2.8.0
python-dotenv==1.0.1
requests==2.32.3
resend==2.1.0
rest-framework-simplejwt==0.0.2
sniffio==1.3.1
sqlparse==0.5.0
stripe==4.2.0
tqdm==4.66.4
typing_extensions==4.12.1
urllib3==2.2.1
yarl==1.9.4
